Created utils/classification and utils/residual_stream
