+++ fit_directions.py
@@ -36,6 +36,8 @@ from transformer_lens.hook_points import (
 import wandb
 from utils.store import save_array, save_html, update_csv, get_csv, eval_csv
 from utils.prompts import get_dataset, filter_words_by_length, PromptType
+from utils.residual_stream import ResidualStreamDataset
+from utils.classification import train_kmeans, CSV_COLS
 #%%
 # ============================================================================ #
 # model loading
@@ -51,371 +53,6 @@ model = HookedTransformer.from_pretrained(
     device=device,
 )
 model.name = MODEL_NAME
-#%%
-# ============================================================================ #
-# Data loading
-CSV_COLS = (
-    'train_set', 'train_pos', 'train_layer', 'test_set', 'test_pos', 'test_layer'
-)
-#%%
-class ResidualStreamDataset:
-
-    def __init__(
-        self, 
-        prompt_strings: List[str], 
-        prompt_tokens: Int[Tensor, "batch pos"], 
-        binary_labels: Int[Tensor, "batch"],
-        model: HookedTransformer,
-        prompt_type: str,
-    ) -> None:
-        assert len(prompt_strings) == len(prompt_tokens)
-        assert len(prompt_strings) == len(binary_labels)
-        self.prompt_strings = prompt_strings
-        self.prompt_tokens = prompt_tokens
-        self.binary_labels = binary_labels
-        self.model = model
-        self.prompt_type = prompt_type
-        example_str_tokens = model.to_str_tokens(prompt_strings[0])
-        self.example = [f"{i}:{tok}" for i, tok in enumerate(example_str_tokens)]
-        self.placeholder_dict = prompt_type.get_placeholder_positions(example_str_tokens)
-        label_positions = [pos for _, positions in self.placeholder_dict.items() for pos in positions]
-        self.str_labels = [
-            ''.join([model.to_str_tokens(prompt)[pos] for pos in label_positions]) 
-            for prompt in prompt_strings
-        ]
-
-    def __len__(self) -> int:
-        return len(self.prompt_strings)
-    
-    def __eq__(self, other: object) -> bool:
-        return set(self.prompt_strings) == set(other.prompt_strings)
-    
-    def embed(self, position_type: str, layer: int) -> Float[Tensor, "batch d_model"]:
-        assert 0 <= layer <= self.model.cfg.n_layers
-        assert position_type in self.placeholder_dict.keys(), (
-            f"Position type {position_type} not found in {self.placeholder_dict.keys()} "
-            f"for prompt type {self.prompt_type}"
-        )
-        embed_position = self.placeholder_dict[position_type][-1]
-        hook = 'resid_pre'
-        if layer == self.model.cfg.n_layers:
-            hook = 'resid_post'
-            layer -= 1
-        _, cache = self.model.run_with_cache(
-            self.prompt_tokens, return_type=None, names_filter = lambda name: hook in name
-        )
-        out: Float[Tensor, "batch pos d_model"] = cache[hook, layer]
-        return out[:, embed_position, :].cpu().detach()
-
-#%%
-def get_km_dataset(
-    model: HookedTransformer,
-    device: torch.device,
-    prompt_type: str = "simple_train",
-    position_type: str = 'ADJ',
-) -> ResidualStreamDataset:
-    all_prompts, answer_tokens, clean_tokens, _ = get_dataset(model, device, prompt_type=prompt_type)
-    clean_labels = answer_tokens[:, 0, 0] == answer_tokens[0, 0, 0]
-    
-    assert len(all_prompts) == len(answer_tokens)
-    assert len(all_prompts) == len(clean_tokens)
-    return ResidualStreamDataset(
-        all_prompts,
-        clean_tokens,
-        clean_labels,
-        position_type,
-        model,
-        prompt_type,
-    )
-#%% # helper functions
-def split_by_label(
-    adjectives: List[str], labels: Int[np.ndarray, "batch"]
-) -> Tuple[List[str], List[str]]:
-    first_cluster = [
-        adj[1:]
-        for i, adj in enumerate(adjectives) 
-        if labels[i] == 0
-    ]
-    second_cluster = [
-        adj[1:]
-        for i, adj in enumerate(adjectives) 
-        if labels[i] == 1
-    ]
-    return first_cluster, second_cluster
-
-
-def get_accuracy(
-    predicted_positive: Iterable[str],
-    predicted_negative: Iterable[str],
-    actual_positive: Iterable[str],
-    actual_negative: Iterable[str],
-) -> Tuple[str, int, int, float]:
-    correct = (
-        len(set(predicted_positive) & set(actual_positive)) +
-        len(set(predicted_negative) & set(actual_negative))
-    )
-    total = len(actual_positive) + len(actual_negative)
-    accuracy = correct / total
-    return correct, total, accuracy
-
-#%%
-# ============================================================================ #
-# K-means
-def _fit_kmeans(
-    train_data: ResidualStreamDataset, train_pos: str, train_layer: int,
-    test_data: ResidualStreamDataset, test_pos: str, test_layer: int,
-    n_init: int = 10,
-    n_clusters: int = 2,
-    random_state: int = 0,
-):
-    train_embeddings = train_data.embed(train_pos, train_layer)
-    test_embeddings = test_data.embed(train_pos, test_layer)
-    train_positive_adjectives = [
-        adj.strip() for adj, label in zip(train_data.str_labels, train_data.binary_labels) if label == 1
-    ]
-    train_negative_adjectives = [
-        adj.strip() for adj, label in zip(train_data.str_labels, train_data.binary_labels) if label == 0
-    ]
-    test_positive_adjectives = [
-        adj.strip() for adj, label in zip(test_data.str_labels, test_data.binary_labels) if label == 1
-    ]
-    test_negative_adjectives = [
-        adj.strip() for adj, label in zip(test_data.str_labels, test_data.binary_labels) if label == 0
-    ]
-    kmeans = KMeans(n_clusters=n_clusters, n_init=n_init, random_state=random_state)
-    kmeans.fit(train_embeddings)
-    train_km_labels: Int[np.ndarray, "batch"] = kmeans.labels_
-    test_km_labels = kmeans.predict(test_embeddings)
-    km_centroids: Float[np.ndarray, "cluster d_model"] = kmeans.cluster_centers_
-
-    km_first_cluster, km_second_cluster = split_by_label(
-        train_data.str_labels, train_km_labels
-    )
-    one_pos = len(set(km_first_cluster) & set(train_positive_adjectives))
-    one_neg = len(set(km_first_cluster) & set(train_negative_adjectives))
-    two_pos = len(set(km_second_cluster) & set(train_positive_adjectives))
-    two_neg = len(set(km_second_cluster) & set(train_negative_adjectives))
-    pos_first = one_pos + two_neg > one_neg + two_pos
-    if pos_first:
-        train_positive_cluster = km_first_cluster
-        train_negative_cluster = km_second_cluster
-        km_positive_centroid = km_centroids[0, :]
-        km_negative_centroid = km_centroids[1, :]
-        test_positive_cluster, test_negative_cluster = split_by_label(
-            test_data.str_labels, test_km_labels
-        )
-    else:
-        train_positive_cluster = km_second_cluster
-        train_negative_cluster = km_first_cluster
-        km_positive_centroid = km_centroids[1, :]
-        km_negative_centroid = km_centroids[0, :]
-        test_negative_cluster, test_positive_cluster = split_by_label(
-            test_data.str_labels, test_km_labels
-        )
-        train_km_labels = 1 - train_km_labels
-        test_km_labels = 1 - test_km_labels
-    km_line: Float[np.ndarray, "d_model"] = (
-        km_positive_centroid - km_negative_centroid
-    )
-    # get accuracy
-    _, _, insample_accuracy = get_accuracy(
-        train_positive_cluster,
-        train_negative_cluster,
-        train_positive_adjectives,
-        train_negative_adjectives,
-    )
-    correct, total, accuracy = get_accuracy(
-        test_positive_cluster,
-        test_negative_cluster,
-        test_positive_adjectives,
-        test_negative_adjectives,
-    )
-    is_insample = (
-        train_data.prompt_type == test_data.prompt_type and
-        train_pos == test_pos and
-        train_layer == test_layer
-    )
-    if is_insample:
-        assert accuracy >= 0.5, (
-            f"Accuracy should be at least 50%, got {accuracy:.1%}, "
-            f"direct calc:{insample_accuracy:.1%}, "
-            f"train:{train_data.prompt_type.value}, layer:{train_layer}, \n"
-            f"positive cluster: {sorted(test_positive_cluster)}\n"
-            f"negative cluster: {sorted(test_negative_cluster)}\n"
-            f"positive adjectives: {sorted(test_positive_adjectives)}\n"
-            f"negative adjectives: {sorted(test_negative_adjectives)}\n"
-        )
-    
-    return km_line, correct, total, accuracy
-#%%
-def safe_cosine_sim(
-    line1: Float[np.ndarray, "d_model"], line2: Float[np.ndarray, "d_model"],
-    tol: float = 1e-6, min_value: float = -0.9,
-):
-    if np.linalg.norm(line1) < tol or np.linalg.norm(line2) < tol:
-        cosine_sim = 0
-    else:
-        cosine_sim = np.dot(line1, line2) / (
-            np.linalg.norm(line1) * np.linalg.norm(line2)
-        )
-    assert cosine_sim >= min_value, (
-        f"cosine sim very negative ({cosine_sim:.2f}), looks like a flipped sign"
-    )
-    return cosine_sim
-#%%
-def train_kmeans(
-    train_data: ResidualStreamDataset, train_pos: str, train_layer: int,
-    test_data: ResidualStreamDataset, test_pos: str, test_layer: int,
-    n_init: int = 10,
-    n_clusters: int = 2,
-    random_state: int = 0,
-):
-    km_line, correct, total, accuracy = _fit_kmeans(
-        train_data, train_pos, train_layer,
-        test_data, test_pos, test_layer,
-        n_init=n_init,
-        n_clusters=n_clusters,
-        random_state=random_state,
-    )
-    test_line, _, _, _ = _fit_kmeans(
-        test_data, test_layer,
-        test_data, test_layer,
-        n_init=n_init,
-        n_clusters=n_clusters,
-        random_state=random_state,
-    )
-    # write k means line to file
-    save_array(km_line, f"km_{train_data.prompt_type.value}_{train_pos}_layer{train_layer}", model)
-
-    cosine_sim = safe_cosine_sim(km_line, test_line)
-    columns = [
-        'train_set', 'train_layer', 'train_pos',
-        'test_set', 'test_layer',  'test_pos',
-        'correct', 'total', 'accuracy', 'similarity',
-    ]
-    data = [[
-        train_data.prompt_type.value, train_layer, train_pos,
-        test_data.prompt_type.value, test_layer, test_pos,
-        correct, total, accuracy, cosine_sim,
-    ]]
-    stats_df = pd.DataFrame(data, columns=columns)
-    
-    update_csv(
-        stats_df, "km_stats", model, key_cols=CSV_COLS
-    )
-
-# %%
-# ============================================================================ #
-# PCA
-def train_pca(
-    train_data: ResidualStreamDataset, train_pos: str, train_layer: int,
-    test_data: ResidualStreamDataset, test_pos: str, test_layer: int,
-    n_init: int = 10,
-    n_clusters: int = 2,
-    random_state: int = 0,
-) -> Tuple[pd.DataFrame, go.Figure]:
-    train_embeddings = train_data.embed(train_pos, train_layer)
-    test_embeddings = test_data.embed(test_pos, test_layer)
-    train_positive_adjectives = [
-        adj.strip() for adj, label in zip(train_data.str_labels, train_data.binary_labels) if label == 1
-    ]
-    train_negative_adjectives = [
-        adj.strip() for adj, label in zip(train_data.str_labels, train_data.binary_labels) if label == 0
-    ]
-    test_positive_adjectives = [
-        adj.strip() for adj, label in zip(test_data.str_labels, test_data.binary_labels) if label == 1
-    ]
-    test_negative_adjectives = [
-        adj.strip() for adj, label in zip(test_data.str_labels, test_data.binary_labels) if label == 0
-    ]
-    pca = PCA(n_components=2)
-    train_pcs = pca.fit_transform(train_embeddings.numpy())
-    test_pcs = pca.transform(test_embeddings.numpy())
-    kmeans = KMeans(n_clusters=n_clusters, n_init=n_init, random_state=random_state)
-    kmeans.fit(train_pcs)
-    train_pca_labels: Int[np.ndarray, "batch"] = kmeans.labels_
-    test_pca_labels = kmeans.predict(test_pcs)
-    pca_centroids: Float[np.ndarray, "cluster pca"] = kmeans.cluster_centers_
-    # PCA components should already be normalised, but just in case
-    comp_unit = pca.components_[0, :] / np.linalg.norm(pca.components_[0, :])
-    save_array(comp_unit, f"pca_0_{train_data.prompt_type.value}_layer{train_layer}", model)
-    pca_first_cluster, pca_second_cluster = split_by_label(
-        train_data.str_labels, train_pca_labels
-    )
-    one_pos = len(set(pca_first_cluster) & set(train_positive_adjectives))
-    one_neg = len(set(pca_first_cluster) & set(train_negative_adjectives))
-    two_pos = len(set(pca_second_cluster) & set(train_positive_adjectives))
-    two_neg = len(set(pca_second_cluster) & set(train_negative_adjectives))
-    pca_pos_first = one_pos + two_neg > one_neg + two_pos
-    if pca_pos_first:
-        # positive first
-        train_pca_positive_cluster = pca_first_cluster
-        train_pca_negative_cluster = pca_second_cluster
-        test_pca_positive_cluster, test_pca_negative_cluster = split_by_label(
-            test_data.str_labels, test_pca_labels
-        )
-        pca_positive_centroid = pca_centroids[0, :]
-        pca_negative_centroid = pca_centroids[1, :]
-    else:
-        # negative first
-        train_pca_positive_cluster = pca_second_cluster
-        train_pca_negative_cluster = pca_first_cluster
-        test_pca_negative_cluster, test_pca_positive_cluster = split_by_label(
-            test_data.str_labels, test_pca_labels
-        )
-        pca_negative_centroid = pca_centroids[0, :]
-        pca_positive_centroid = pca_centroids[1, :]
-        train_pca_labels = 1 - train_pca_labels
-        test_pca_labels = 1 - test_pca_labels
-    
-    test_pca = PCA(n_components=2)
-    test_pca.fit_transform(train_embeddings.numpy())
-    test_comp_unit = test_pca.components_[0, :] / np.linalg.norm(test_pca.components_[0, :])
-    # compute cosine sim
-    if np.linalg.norm(test_comp_unit) < 1e-6:
-        cosine_sim = 0
-    else:
-        cosine_sim = np.dot(comp_unit, test_comp_unit) / (
-            np.linalg.norm(comp_unit) * np.linalg.norm(test_comp_unit)
-        )
-    correct, total, accuracy = get_accuracy(
-        test_pca_positive_cluster,
-        test_pca_negative_cluster,
-        test_positive_adjectives,
-        test_negative_adjectives,
-    )
-    columns = [
-        'train_set', 'train_layer', 'train_pos',
-        'test_set', 'test_layer',  'test_pos',
-        'correct', 'total', 'accuracy', 'similarity',
-    ]
-    data = [[
-        train_data.prompt_type.value, train_layer, train_pos,
-        test_data.prompt_type.value, test_layer, test_pos,
-        correct, total, accuracy, cosine_sim,
-    ]]
-    stats_df = pd.DataFrame(data, columns=columns)
-    update_csv(
-        stats_df, "pca_stats", model, key_cols=CSV_COLS
-    )
-    plot_data = [[
-         train_data.prompt_type.value, train_layer, train_pos,
-        test_data.prompt_type.value, test_layer, test_pos,
-        train_pcs, train_data.str_labels, train_data.binary_labels, 
-        test_pcs, test_data.str_labels, test_data.binary_labels,
-        pca_centroids
-    ]]
-    plot_columns = [
-        'train_set', 'train_layer', 'train_pos',
-        'test_set', 'test_layer',  'test_pos',
-        'train_pcs', 'train_str_labels', 'train_true_labels',
-        'test_pcs', 'test_str_labels', 'test_true_labels',
-        'pca_centroids',
-    ]
-    plot_df = pd.DataFrame(plot_data, columns=plot_columns)
-    update_csv(
-        plot_df, "pca_plot", model, key_cols=CSV_COLS
-    )
 #%%
 PROMPT_TYPES = [
     PromptType.SIMPLE_TRAIN,
@@ -454,19 +91,21 @@ for train_type, train_layer, test_type, test_layer in BAR:
         # if eval_csv(query, "km_stats", model):
         #     continue
 
-        trainset = get_km_dataset(model, device, prompt_type=train_type, position_type=train_pos)
-        testset = get_km_dataset(model, device, prompt_type=test_type, position_type=test_pos)
+        trainset = ResidualStreamDataset.get_dataset(model, device, prompt_type=train_type)
+        testset = ResidualStreamDataset.get_dataset(model, device, prompt_type=test_type)
         train_kmeans(
             trainset, train_layer,
             testset, test_layer,
         )
-        train_pca(
+        train_kmeans(
             trainset, train_layer,
             testset, test_layer,
+            pca_components=2,
         )
 #%%
 km_stats = get_csv("km_stats", model, key_cols=CSV_COLS)
 km_stats = km_stats.loc[
+    km_stats.method.eq('kmeans')
     km_stats.train_set.isin(['simple_train']) & 
     km_stats.train_pos.isin(['ADJ']) &
     km_stats.test_set.isin(['simple_train', 'simple_test', 'simple_adverb'])
